Make sure HarmonicNetworkCore gets around
