+++ Sources/HarmonicNetworkCore/ChordCluster.swift
@@ -10,3 +10,4 @@ import Pitch
 import SpelledPitch
 
 public typealias ChordClusterNode = Tree<String,String>
+
